️🔨create resource-server module
