+++ resource-server/src/main/java/pe/goblin/resourceserver/ResourceserverApplication.java
@@ -1,0 +1,13 @@
+package pe.goblin.resourceserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ResourceserverApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ResourceserverApplication.class, args);
+    }
+
+}
