+++ resource-server/src/main/java/pe/goblin/resourceserver/domain/support/mail/dto/Status.java
@@ -1,0 +1,5 @@
+package pe.goblin.resourceserver.domain.support.mail.dto;
+
+public enum Status {
+    REQUESTED, SENDING, SUCCESS, FAILED
+}
