+++ resource-server/src/main/java/pe/goblin/resourceserver/domain/support/mail/exception/MailException.java
@@ -1,0 +1,11 @@
+package pe.goblin.resourceserver.domain.support.mail.exception;
+
+public abstract class MailException extends RuntimeException {
+    protected MailException(String message) {
+        super(message);
+    }
+
+    protected MailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
