+++ resource-server/src/main/java/pe/goblin/resourceserver/domain/support/mail/exception/MailTemplateException.java
@@ -1,0 +1,12 @@
+package pe.goblin.resourceserver.domain.support.mail.exception;
+
+public class MailTemplateException extends MailException {
+
+    public MailTemplateException(String message) {
+        super(message);
+    }
+
+    public MailTemplateException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
