+++ resource-server/src/main/resources/application.yaml
