+++ resource-server/src/main/resources/mapper/mail_event_repository.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="pe.goblin.resourceserver.domain.support.mail.repository.MyBatisMailEventRepository">
+
+    <insert id="save" useGeneratedKeys="true" keyProperty="id" keyColumn="id" flushCache="true">
+        INSERT INTO mail_event (receiver, subject, template, arguments, status, created_at)
+        VALUES (#{receiver}, #{subject}, #{template},
+        #{arguments, typeHandler=pe.goblin.resourceserver.common.mybatis.typehandler.JsonToMapTypeHandler}, #{status},
+        NOW())
+    </insert>
+
+    <insert id="saveAll" useGeneratedKeys="true" keyProperty="id" keyColumn="id" flushCache="true">
+        INSERT INTO mail_event (receiver, subject, template, arguments, status, created_at)
+        VALUES
+        <foreach collection="list" item="mailEvent" separator=",">
+            (
+            #{mailEvent.receiver},
+            #{mailEvent.subject},
+            #{mailEvent.template},
+            #{mailEvent.arguments, typeHandler=pe.goblin.resourceserver.common.mybatis.typehandler.JsonToMapTypeHandler},
+            #{mailEvent.status},
+            NOW()
+            )
+        </foreach>
+    </insert>
+
+    <select id="findById" resultMap="MailEventResultMap" useCache="false">
+        SELECT * FROM mail_event WHERE id = #{id}
+    </select>
+
+    <select id="findAllById" resultMap="MailEventResultMap" useCache="false">
+        SELECT * FROM mail_event WHERE id IN
+        <foreach collection="ids" item="id" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+    </select>
+
+    <select id="findAllByStatus" resultMap="MailEventResultMap" useCache="false">
+        SELECT * FROM mail_event WHERE status = #{status} ORDER BY created_at ASC LIMIT #{limit}
+    </select>
+
+    <update id="updateStatus" flushCache="true">
+        UPDATE mail_event
+        SET status = #{status}, updated_at = NOW()
+        WHERE id = #{id}
+    </update>
+
+    <update id="updateStatusInBulk" flushCache="true">
+        UPDATE mail_event
+        SET status = #{status}, updated_at = NOW()
+        WHERE id IN
+        <foreach item="id" collection="ids" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+    </update>
+
+    <resultMap id="MailEventResultMap" type="pe.goblin.resourceserver.domain.support.mail.dto.MailEvent">
+        <id property="id" column="id"/>
+        <result property="receiver" column="receiver"/>
+        <result property="subject" column="subject"/>
+        <result property="template" column="template"
+                javaType="pe.goblin.resourceserver.domain.support.mail.dto.MailTemplate"
+                typeHandler="org.apache.ibatis.type.EnumTypeHandler"/>
+        <result property="arguments" column="arguments" javaType="java.util.Map"
+                typeHandler="pe.goblin.resourceserver.common.mybatis.typehandler.JsonToMapTypeHandler"/>
+        <result property="status" column="status" javaType="pe.goblin.resourceserver.domain.support.mail.dto.Status"
+                typeHandler="org.apache.ibatis.type.EnumTypeHandler"/>
+        <result property="createdAt" column="created_at"/>
+        <result property="updatedAt" column="updated_at"/>
+    </resultMap>
+</mapper>
