+++ resource-server/src/main/resources/templates/mails/mail-verification.html
