+++ resource-server/src/test/java/pe/goblin/resourceserver/ResourceserverApplicationTests.java
@@ -1,0 +1,13 @@
+package pe.goblin.resourceserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ResourceserverApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
