+++ settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'itda'
+include 'resource-server'
+
