--- src/main/java/pe/goblin/itda/ItdaApplication.java
@@ -1,13 +1,0 @@
-package pe.goblin.itda;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ItdaApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ItdaApplication.class, args);
-    }
-
-}
