--- src/main/java/pe/goblin/itda/common/mybatis/typehandler/JsonToMapTypeHandler.java
@@ -1,51 +1,0 @@
-package pe.goblin.itda.common.mybatis.typehandler;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.ibatis.type.BaseTypeHandler;
-import org.apache.ibatis.type.JdbcType;
-
-import java.sql.CallableStatement;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.Map;
-
-public class JsonToMapTypeHandler extends BaseTypeHandler<Map<String, Object>> {
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    @Override
-    public void setNonNullParameter(PreparedStatement ps, int i, Map<String, Object> parameter, JdbcType jdbcType) throws SQLException {
-        try {
-            ps.setString(i, objectMapper.writeValueAsString(parameter));
-        } catch (JsonProcessingException e) {
-            throw new SQLException("Failed to convert Map to JSON", e);
-        }
-    }
-
-    @Override
-    public Map<String, Object> getNullableResult(ResultSet rs, String columnName) throws SQLException {
-        return parseJson(rs.getString(columnName));
-    }
-
-    @Override
-    public Map<String, Object> getNullableResult(ResultSet rs, int columnIndex) throws SQLException {
-        return parseJson(rs.getString(columnIndex));
-    }
-
-    @Override
-    public Map<String, Object> getNullableResult(CallableStatement cs, int columnIndex) throws SQLException {
-        return parseJson(cs.getString(columnIndex));
-    }
-
-    private Map<String, Object> parseJson(String json) throws SQLException {
-        try {
-            return json == null ? new HashMap<>() : objectMapper.readValue(json, new TypeReference<>() {
-            });
-        } catch (JsonProcessingException e) {
-            throw new SQLException("Error parsing JSON to Map", e);
-        }
-    }
-}
