--- src/main/java/pe/goblin/itda/domain/support/mail/dto/Status.java
@@ -1,5 +1,0 @@
-package pe.goblin.itda.domain.support.mail.dto;
-
-public enum Status {
-    REQUESTED, SENDING, SUCCESS, FAILED
-}
