--- src/main/java/pe/goblin/itda/domain/support/mail/exception/MailException.java
@@ -1,11 +1,0 @@
-package pe.goblin.itda.domain.support.mail.exception;
-
-public abstract class MailException extends RuntimeException {
-    protected MailException(String message) {
-        super(message);
-    }
-
-    protected MailException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
