--- src/main/java/pe/goblin/itda/domain/support/mail/exception/MailTemplateException.java
@@ -1,12 +1,0 @@
-package pe.goblin.itda.domain.support.mail.exception;
-
-public class MailTemplateException extends MailException {
-
-    public MailTemplateException(String message) {
-        super(message);
-    }
-
-    public MailTemplateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
