--- src/test/java/pe/goblin/itda/ItdaApplicationTests.java
@@ -1,13 +1,0 @@
-package pe.goblin.itda;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ItdaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
